Mode has now a fixed int as key
provided in mode.json
may be temporary

+++ Android/app/src/main/kotlin/com/onewisebit/scpescape/model/daos/ModeDAO.kt
@@ -14,21 +14,21 @@ interface ModeDAO {
      * @return the Mode from the table with a specific id.
      */
     @Query("SELECT * FROM modes WHERE mode_ID = :id")
-    fun getModeById(id: Long): Mode
+    fun getModeById(id: Int): Mode
 
     /**
      * Get the minimum number of players from game mode id.
      * @return the minimum number of players needed to play this mode.
      */
     @Query("SELECT min_players FROM modes WHERE mode_ID = :id")
-    fun getMinPlayers(id: Long): Int
+    fun getMinPlayers(id: Int): Int
 
     /**
      * Get the maximum number of players from game mode id.
      * @return the maximal number of players needed to play this mode.
      */
     @Query("SELECT max_players FROM modes WHERE mode_ID = :id")
-    fun getMaxPlayers(id: Long): Int
+    fun getMaxPlayers(id: Int): Int
 
     /**
      * Insert a Mode in the database. If the mode already exists, replace it.
